# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Thesis Proposal
  version: 1.0.0
servers:
- url: /
...
paths:
  /professors:
    get:
      tags:
      - professor
      operationId: getProfessors
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/professors'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Professor
  /professors/{professorId}:
    get:
      tags:
      - professor
      operationId: getProfessorById
      parameters:
      - name: professorId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/professor'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Professor
...
  /thesisProposals:
    get:
      tags:
      - thesisProposal
      operationId: getThesisProposals
      parameters:
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the professor that perform
          this request
        required: false
        style: form
        explode: true
        schema:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
      - name: cosupervisor
        in: query
        description: Respresent the filter for the professor
        required: false
        style: form
        explode: true
        schema:
          type: boolean
      - name: codDegree
        in: cookie
        description: "The codDegree is about the degree that the student, that perform\
          \ this request, attends"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: keywords
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: supervisor
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: title
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: thesisType
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: abroad
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: boolean
      - name: expirationDate
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/thesisProposals'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: ThesisProposal
    post:
      tags:
      - thesisProposal
      operationId: insertNewThesisProposal
      parameters:
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the professor that perform
          this request
        required: true
        style: form
        explode: true
        schema:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/thesisProposal'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/thesisProposal'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: ThesisProposal
  /thesisProposals/{thesisProposalId}:
    get:
      tags:
      - thesisProposal
      operationId: getThesisProposalById
      parameters:
      - name: thesisProposalId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/thesisProposal'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: ThesisProposal
    post:
      tags:
      - application
      operationId: insertNewApplication
      parameters:
      - name: thesisProposalId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the student that perform
          this request
        required: true
        style: form
        explode: true
        schema:
          pattern: "^(s|S)[0-9]{6}$"
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/application'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/application'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Application
  /applications:
    get:
      tags:
      - application
      operationId: getApplications
      parameters:
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the professor that perform
          this request
        required: true
        style: form
        explode: true
        schema:
          pattern: "^((p|P)[0-9]{6})|((s|S)[0-9]{6})$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/applications'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Application
  /applications/{thesisProposalId}/{studentId}:
    get:
      tags:
      - application
      operationId: getApplicationById
      parameters:
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the professor that perform
          this request
        required: true
        style: form
        explode: true
        schema:
          pattern: "^((p|P)[0-9]{6})|((s|S)[0-9]{6})$"
          type: string
      - name: thesisProposalId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: studentId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/application'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Application
    put:
      tags:
      - application
      operationId: updateApplication
      parameters:
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the professor that perform
          this request
        required: true
        style: form
        explode: true
        schema:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
      - name: thesisProposalId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: studentId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/application'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/application'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Application
  /externalCoSupervisors:
    get:
      tags:
      - externalCoSupervisor
      operationId: getExternalCoSupervisors
      parameters:
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the professor that perform
          this request
        required: true
        style: form
        explode: true
        schema:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/externalCoSupervisors'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      x-swagger-router-controller: ExternalCoSupervisor
  /externalCoSupervisors/{externalCoSupervisorId}:
    get:
      tags:
      - externalCoSupervisor
      operationId: getExternalCoSupervisorById
      parameters:
      - name: authenticatedUserId
        in: cookie
        description: The authenticated user id corresponds to the professor that perform
          this request
        required: true
        style: form
        explode: true
        schema:
          pattern: "^((p|P)|(s|S))[0-9]{6}$"
          type: string
      - name: externalCoSupervisorId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          pattern: "^(e|E)[0-9]{6}$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/externalCoSupervisor'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      x-swagger-router-controller: ExternalCoSupervisor
...
components:
  schemas:
...
    professor:
      required:
      - email
      type: object
      properties:
        professorId:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
        name:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        surname:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        email:
          pattern: "^[a-z]+\\.[a-z]+@polito\\.it$"
          type: string
          format: email
        codGroup:
          pattern: "^CG_[0-9]{2}$"
          type: string
        codDepartment:
          pattern: "^CD_[0-9]{2}$"
          type: string
        self:
          type: string
          format: uri
      additionalProperties: false
      example:
        codDepartment: codDepartment
        surname: surname
        name: name
        self: http://example.com/aeiou
        professorId: professorId
        codGroup: codGroup
        email: ""
    professors:
      type: array
      items:
        $ref: '#/components/schemas/professors_inner'
...
    application:
      required:
      - thesisProposalId
      type: object
      properties:
        thesisProposalId:
          type: integer
        thesisProposalTitle:
          minLength: 1
          type: string
        applicant:
          $ref: '#/components/schemas/application_applicant'
        message:
          maxLength: 1000
          type: string
        date:
          type: string
          format: date
        isReadedByProfessor:
          type: boolean
          default: false
        isReadedByStudent:
          type: boolean
          default: false
        isAccepted:
          type: string
          default: Pending
          enum:
          - Rejected
          - Pending
          - Accepted
      additionalProperties: false
      example:
        date: 2000-01-23
        isReadedByProfessor: false
        thesisProposalTitle: thesisProposalTitle
        isAccepted: ""
        message: message
        thesisProposalId: 0
        applicant:
          studentId: studentId
          student: http://example.com/aeiou
          surname: surname
          name: name
        isReadedByStudent: false
    applications:
      type: array
      items:
        $ref: '#/components/schemas/applications_inner'
    thesisProposal:
      required:
      - CdS
      - description
      - expirationDate
      - keywords
      - level
      - requirements
      - thesisType
      - title
      type: object
      properties:
        thesisProposalId:
          minimum: 1
          type: integer
        title:
          minLength: 1
          type: string
        supervisor:
          $ref: '#/components/schemas/thesisProposal_supervisor'
        coSupervisor:
          type: array
          items:
            $ref: '#/components/schemas/thesisProposal_coSupervisor'
        keywords:
          minItems: 1
          type: array
          items:
            type: string
        thesisType:
          minLength: 1
          type: string
        abroad:
          type: boolean
        groups:
          minItems: 1
          type: array
          items:
            minLength: 1
            type: string
        description:
          maxLength: 1000
          minLength: 1
          type: string
        requirements:
          maxLength: 1000
          minLength: 1
          type: string
        notes:
          maxLength: 1000
          minLength: 1
          type: string
        expirationDate:
          type: string
          format: date
        level:
          type: string
          enum:
          - BSc
          - MSc
        CdS:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/thesisProposal_CdS'
        isArchieved:
          type: boolean
          default: false
      additionalProperties: false
      example:
        abroad: true
        requirements: requirements
        notes: notes
        keywords:
        - keywords
        - keywords
        level: ""
        thesisType: thesisType
        groups:
        - groups
        - groups
        description: description
        title: title
        coSupervisor:
        - coSupervisorId: coSupervisorId
          coSupervisor: http://example.com/aeiou
          surname: surname
          name: name
        - coSupervisorId: coSupervisorId
          coSupervisor: http://example.com/aeiou
          surname: surname
          name: name
        isArchieved: false
        CdS:
        - degreeId: degreeId
          titleDegree: titleDegree
        - degreeId: degreeId
          titleDegree: titleDegree
        supervisor:
          professor: http://example.com/aeiou
          surname: surname
          name: name
          professorId: professorId
        thesisProposalId: 1
        expirationDate: 2000-01-23
    thesisProposals:
      type: array
      items:
        $ref: '#/components/schemas/thesisProposals_inner'
    externalCoSupervisor:
      type: object
      properties:
        externalCoSupervisorId:
          pattern: "^(e|E)[0-9]{6}$"
          type: string
        name:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        surname:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        email:
          type: string
          format: email
        company:
          minLength: 1
          type: string
        self:
          type: string
          format: uri
      additionalProperties: false
      example:
        surname: surname
        name: name
        externalCoSupervisorId: externalCoSupervisorId
        self: http://example.com/aeiou
        company: company
        email: ""
    externalCoSupervisors:
      type: array
      items:
        $ref: '#/components/schemas/externalCoSupervisors_inner'
...
    professors_inner:
      type: object
      properties:
        professorId:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
        name:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        surname:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        self:
          type: string
          format: uri
      additionalProperties: false
      example:
        surname: surname
        name: name
        self: http://example.com/aeiou
        professorId: professorId
    application_applicant:
      type: object
      properties:
        studentId:
          pattern: "^(s|S)[0-9]{6}$"
          type: string
        name:
          minLength: 1
          type: string
        surname:
          minLength: 1
          type: string
        student:
          type: string
          format: uri
      additionalProperties: false
      example:
        studentId: studentId
        student: http://example.com/aeiou
        surname: surname
        name: name
    applications_inner:
      type: object
      properties:
        thesisProposalId:
          minimum: 1
          type: integer
        thesisProposalTitle:
          minLength: 1
          type: string
        studentId:
          pattern: "^(s|S)[0-9]{6}$"
          type: string
        date:
          type: string
          format: date
        self:
          type: string
          format: uri
      additionalProperties: false
      example:
        studentId: studentId
        date: 2000-01-23
        thesisProposalTitle: thesisProposalTitle
        self: http://example.com/aeiou
        thesisProposalId: 1
    thesisProposal_supervisor:
      type: object
      properties:
        professorId:
          pattern: "^(p|P)[0-9]{6}$"
          type: string
        name:
          minLength: 1
          type: string
        surname:
          minLength: 1
          type: string
        professor:
          type: string
          format: uri
      additionalProperties: false
      example:
        professor: http://example.com/aeiou
        surname: surname
        name: name
        professorId: professorId
    thesisProposal_coSupervisor:
      required:
      - coSupervisorId
      type: object
      properties:
        coSupervisorId:
          pattern: "^((p|P)|(e|E))[0-9]{6}$"
          type: string
        name:
          minLength: 1
          type: string
        surname:
          minLength: 1
          type: string
        coSupervisor:
          type: string
          format: uri
      additionalProperties: false
      example:
        coSupervisorId: coSupervisorId
        coSupervisor: http://example.com/aeiou
        surname: surname
        name: name
    thesisProposal_CdS:
      required:
      - degreeId
      type: object
      properties:
        degreeId:
          pattern: "^LM?-([1-9]|[1-9][0-9])$"
          type: string
        titleDegree:
          minLength: 1
          type: string
      additionalProperties: false
      example:
        degreeId: degreeId
        titleDegree: titleDegree
    thesisProposals_inner:
      type: object
      properties:
        thesisProposalId:
          minimum: 1
          type: integer
        title:
          minLength: 1
          type: string
        name:
          minLength: 1
          type: string
        surname:
          minLength: 1
          type: string
        keywords:
          minItems: 1
          type: array
          items:
            type: string
        expirationDate:
          type: string
          format: date
        self:
          type: string
          format: uri
      additionalProperties: false
      example:
        keywords:
        - keywords
        - keywords
        surname: surname
        name: name
        self: http://example.com/aeiou
        title: title
        thesisProposalId: 1
        expirationDate: 2000-01-23
    externalCoSupervisors_inner:
      type: object
      properties:
        externalCoSupervisorId:
          pattern: "^(e|E)[0-9]{6}$"
          type: string
        name:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        surname:
          minLength: 1
          pattern: "^[a-zA-Z ]+$"
          type: string
        self:
          type: string
          format: uri
      additionalProperties: false
      example:
        surname: surname
        name: name
        externalCoSupervisorId: externalCoSupervisorId
        self: http://example.com/aeiou
...